MouseAxisManipulation
Create a new static class to wrap shared logic between old and new interaction controllers
