+++ src/ScottPlot5/ScottPlot5/Control/StandardActions.cs
@@ -1,10 +1,4 @@
-﻿using System.Data;
-
-namespace ScottPlot.Control;
-
-// TODO: refactor individual actions into their own classes which inherit IControlAction
-
-// NOTE: every action should lock the plot so its actions do not affect concurrent renders
+﻿namespace ScottPlot.Control;
 
 public static class StandardActions
 {
@@ -17,7 +11,7 @@ public static class StandardActions
             double xFrac = locked.X ? 1 : zoomInFraction;
             double yFrac = locked.Y ? 1 : zoomInFraction;
 
-            MouseZoom(control.Plot, xFrac, yFrac, pixel, control.Interaction.ChangeOpposingAxesTogether);
+            Interactivity.MouseAxisManipulation.MouseWheelZoom(control.Plot, xFrac, yFrac, pixel, control.Interaction.ChangeOpposingAxesTogether);
             control.Refresh();
         }
     }
@@ -31,7 +25,7 @@ public static class StandardActions
             double xFrac = locked.X ? 1 : zoomOutFraction;
             double yFrac = locked.Y ? 1 : zoomOutFraction;
 
-            MouseZoom(control.Plot, xFrac, yFrac, pixel, control.Interaction.ChangeOpposingAxesTogether);
+            Interactivity.MouseAxisManipulation.MouseWheelZoom(control.Plot, xFrac, yFrac, pixel, control.Interaction.ChangeOpposingAxesTogether);
             control.Refresh();
         }
     }
@@ -94,43 +88,8 @@ public static class StandardActions
 
             Pixel mouseDown = drag.From;
 
-            float pixelDeltaX = -(mouseNow.X - mouseDown.X);
-            float pixelDeltaY = mouseNow.Y - mouseDown.Y;
-
-            float scaledDeltaX = pixelDeltaX / control.Plot.ScaleFactorF;
-            float scaledDeltaY = pixelDeltaY / control.Plot.ScaleFactorF;
-
-            // restore MouseDown limits
             drag.InitialLimits.Apply(control.Plot);
-
-            IAxis? axisUnderMouse = control.Plot.GetAxis(mouseDown);
-
-            PixelRect dataRect = control.Plot.RenderManager.LastRender.DataRect;
-
-            if (axisUnderMouse is not null)
-            {
-                if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
-                {
-                    control.Plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.PanMouse(scaledDeltaX, dataRect.Width));
-                }
-                else if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsVertical())
-                {
-                    control.Plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.PanMouse(scaledDeltaY, dataRect.Height));
-                }
-                else
-                {
-                    float scaledDelta = axisUnderMouse.IsHorizontal() ? scaledDeltaX : scaledDeltaY;
-                    float dataSize = axisUnderMouse.IsHorizontal() ? dataRect.Width : dataRect.Height;
-                    axisUnderMouse.Range.PanMouse(scaledDelta, dataSize);
-                }
-            }
-            else
-            {
-                // modify all axes
-                control.Plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.PanMouse(scaledDeltaX, dataRect.Width));
-                control.Plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.PanMouse(scaledDeltaY, dataRect.Height));
-            }
-
+            Interactivity.MouseAxisManipulation.DragPan(control.Plot, mouseDown, mouseNow);
             control.Refresh();
         }
     }
@@ -145,41 +104,9 @@ public static class StandardActions
 
             Pixel mouseDown = drag.From;
 
-            float pixelDeltaX = mouseNow.X - mouseDown.X;
-            float pixelDeltaY = -(mouseNow.Y - mouseDown.Y);
-
             // restore MouseDown limits
             drag.InitialLimits.Apply(control.Plot);
-
-            IAxis? axisUnderMouse = control.Plot.GetAxis(mouseDown);
-
-            PixelRect lastRenderDataRect = control.Plot.RenderManager.LastRender.DataRect;
-
-            if (axisUnderMouse is not null)
-            {
-                if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
-                {
-                    control.Plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.ZoomMouseDelta(pixelDeltaX, lastRenderDataRect.Width));
-                }
-                else if (control.Interaction.ChangeOpposingAxesTogether && axisUnderMouse.IsVertical())
-                {
-                    control.Plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.ZoomMouseDelta(pixelDeltaY, lastRenderDataRect.Height));
-                }
-                else
-                {
-                    // modify a single axis
-                    float pixelDelta = axisUnderMouse.IsHorizontal() ? pixelDeltaX : pixelDeltaY;
-                    float dataSize = axisUnderMouse.IsHorizontal() ? lastRenderDataRect.Width : lastRenderDataRect.Height;
-                    axisUnderMouse.Range.ZoomMouseDelta(pixelDelta, dataSize);
-                }
-            }
-            else
-            {
-                // modify all axes
-                control.Plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.ZoomMouseDelta(pixelDeltaX, lastRenderDataRect.Width));
-                control.Plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.ZoomMouseDelta(pixelDeltaY, lastRenderDataRect.Height));
-            }
-
+            Interactivity.MouseAxisManipulation.DragZoom(control.Plot, mouseDown, mouseNow);
             control.Refresh();
         }
     }
@@ -315,41 +242,5 @@ public static class StandardActions
     public static void ShowContextMenu(IPlotControl control, Pixel position)
     {
         control.ShowContextMenu(position);
-    }
-
-    private static void MouseZoom(Plot plot, double fracX, double fracY, Pixel pixel, bool ChangeOpposingAxesTogether)
-    {
-        Pixel scaledPixel = pixel.Divide(plot.ScaleFactorF);
-        MultiAxisLimitManager originalLimits = new(plot);
-        PixelRect dataRect = plot.RenderManager.LastRender.DataRect;
-
-        // restore MouseDown limits
-        originalLimits.Apply(plot);
-
-        var axisUnderMouse = plot.GetAxis(scaledPixel);
-
-        if (axisUnderMouse is not null)
-        {
-            if (ChangeOpposingAxesTogether && axisUnderMouse.IsHorizontal())
-            {
-                plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.ZoomFrac(fracX, xAxis.GetCoordinate(scaledPixel.X, dataRect)));
-            }
-            if (ChangeOpposingAxesTogether && axisUnderMouse.IsVertical())
-            {
-                plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.ZoomFrac(fracY, yAxis.GetCoordinate(scaledPixel.Y, dataRect)));
-            }
-            else
-            {
-                double frac = axisUnderMouse.IsHorizontal() ? fracX : fracY;
-                float scaledCoord = (axisUnderMouse.IsHorizontal() ? scaledPixel.X : scaledPixel.Y);
-                axisUnderMouse.Range.ZoomFrac(frac, axisUnderMouse.GetCoordinate(scaledCoord, dataRect));
-            }
-        }
-        else
-        {
-            // modify all axes
-            plot.Axes.XAxes.ForEach(xAxis => xAxis.Range.ZoomFrac(fracX, xAxis.GetCoordinate(scaledPixel.X, dataRect)));
-            plot.Axes.YAxes.ForEach(yAxis => yAxis.Range.ZoomFrac(fracY, yAxis.GetCoordinate(scaledPixel.Y, dataRect)));
-        }
     }
 }
