+++ src/ScottPlot5/ScottPlot5/Interactivity/UserActionResponses/MouseWheelZoom.cs
@@ -22,18 +22,7 @@ public class MouseWheelZoom(Key horizontalLockKey, Key verticalLockKey) : IUserA
         {
             double xFrac = keys.IsPressed(LockHorizontalKey) ? 1 : 1.15;
             double yFrac = keys.IsPressed(LockVerticalKey) ? 1 : 1.15;
-
-            IAxis? axisUnderMouse = plot.GetAxis(mouseDownInput.Pixel);
-
-            if (axisUnderMouse is not null && ZoomAxisUnderMouse)
-            {
-                plot.Axes.Zoom(axisUnderMouse, mouseDownInput.Pixel, xFrac, yFrac, LockParallelAxes);
-            }
-            else
-            {
-                plot.Axes.Zoom(mouseDownInput.Pixel, xFrac, yFrac);
-            }
-
+            MouseAxisManipulation.MouseWheelZoom(plot, xFrac, yFrac, mouseDownInput.Pixel, LockParallelAxes);
             return new ResponseInfo() { RefreshNeeded = true };
         }
 
@@ -41,18 +30,7 @@ public class MouseWheelZoom(Key horizontalLockKey, Key verticalLockKey) : IUserA
         {
             double xFrac = keys.IsPressed(LockHorizontalKey) ? 1 : 0.85;
             double yFrac = keys.IsPressed(LockVerticalKey) ? 1 : 0.85;
-
-            IAxis? axisUnderMouse = plot.GetAxis(mouseUpInput.Pixel);
-
-            if (axisUnderMouse is not null && ZoomAxisUnderMouse)
-            {
-                plot.Axes.Zoom(axisUnderMouse, mouseUpInput.Pixel, xFrac, yFrac, LockParallelAxes);
-            }
-            else
-            {
-                plot.Axes.Zoom(mouseUpInput.Pixel, xFrac, yFrac);
-            }
-
+            MouseAxisManipulation.MouseWheelZoom(plot, xFrac, yFrac, mouseUpInput.Pixel, LockParallelAxes);
             return new ResponseInfo() { RefreshNeeded = true };
         }
